Release 1.0.0
This is an agent-based model written in Netlogo used as a platform to run computational experiments concerning the eviction of renters in the Phoenix area due in part to the Covid-19 pandemic. To use the simulation one must select the Setup button which will load a map of the Phoenix region and load a representational population of renter households. Then the GO button progresses the simulation on a monthly schedule all the while recording the number of evicted households per month.